+++ src/utils/FileUtils.h
@@ -44,7 +44,7 @@ namespace prm
          * @return
          */
         static bool WritePvcamStack(const void* data, uint16_t imageWidth,
-                                    uint16_t imageHeight, uint16_t imageSize,
+                                    uint16_t imageHeight, uint32_t imageSize,
                                     std::string_view filePath,
                                     uint16_t numImages);
